Update .gitignore
resetting .gitignore text to ensure package-lock.json for users are not ignored when pushed to GitHub. I have only done it for the purpose of sharing a template because the package-lock.json file locks the versions of your dependencies, ensuring that anyone who installs the dependencies using npm install gets the exact same versions that you installed.